Change user sync to use Okta ID (#23)
* change external_id to use the okta id and add a dry-run flag
* make sync flags persistent
* remove okta| prefix
* update dependencies

+++ cmd/sync_users.go
@@ -19,7 +19,11 @@ import (
 // syncUsersCmd governor resources
 var syncUsersCmd = &cobra.Command{
 	Use:   "users",
-	Short: "sync governor and okta resources",
+	Short: "sync okta users into governor",
+	Long: `Performs a one-way user sync from Okta to Governor.
+Users that exist in Okta but not in Governor, will be created. Users that exist in Governor but not in Okta, will be deleted.
+This command is intended for doing an initial load of users. It is strongly recommended that you use the dry-run flag first 
+to see what users would be created/deleted in Governor.`,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return syncUsersToGovernor(cmd.Context())
 	},
@@ -32,7 +36,10 @@ func init() {
 // syncUsersToGovernor syncs users from okta to governor
 func syncUsersToGovernor(ctx context.Context) error {
 	l := logger.Desugar()
-	l.Info("starting sync to governor")
+
+	dryrun := viper.GetBool("sync.dryrun")
+
+	l.Info("starting sync to governor", zap.Bool("dry-run", dryrun))
 
 	oc, err := okta.NewClient(
 		okta.WithLogger(logger.Desugar()),
@@ -69,8 +76,8 @@ func syncUsersToGovernor(ctx context.Context) error {
 		l.Debug("processing okta user", zap.String("okta.user.id", u.Id))
 
 		userType, _ := userType(u)
-		if userType == "contractor" || userType == "serviceuser" {
-			l.Debug("skipping contractor or service user", zap.String("okta.user.id", u.Id))
+		if userType == "serviceuser" {
+			l.Debug("skipping service user", zap.String("okta.user.id", u.Id))
 
 			ignored++
 
@@ -97,8 +104,8 @@ func syncUsersToGovernor(ctx context.Context) error {
 			return nil, err
 		}
 
-		// currently we use pingid here, and this is how external ids live in governor
-		extID := fmt.Sprintf("ping|%s", externalID)
+		// the external id in governor is simply the okta id
+		extID := externalID
 
 		// check if user exists in governor
 		gUsers, err := gc.UsersQuery(ctx, map[string][]string{"external_id": {extID}})
@@ -109,7 +116,7 @@ func syncUsersToGovernor(ctx context.Context) error {
 		l.Debug("got governor users response for external id", zap.Any("governor.users", gUsers))
 
 		if len(gUsers) > 1 {
-			logger.Warn("unexpected user count for pingId",
+			logger.Warn("unexpected user count for external_id",
 				zap.String("external.id", externalID),
 				zap.String("okta.user.id", u.Id),
 				zap.Int("num.governor.users", len(gUsers)),
@@ -134,28 +141,30 @@ func syncUsersToGovernor(ctx context.Context) error {
 			zap.String("okta.user.email", email),
 		)
 
-		gUser, err := gc.CreateUser(ctx, &v1alpha1.UserReq{
-			Email:      email,
-			ExternalID: extID,
-			Name:       fmt.Sprintf("%s %s", first, last),
-		})
-		if err != nil {
-			return nil, err
-		}
+		if !dryrun {
+			gUser, err := gc.CreateUser(ctx, &v1alpha1.UserReq{
+				Email:      email,
+				ExternalID: extID,
+				Name:       fmt.Sprintf("%s %s", first, last),
+			})
+			if err != nil {
+				return nil, err
+			}
 
-		l.Debug("created governor user from okta sync",
-			zap.String("governor.user.id", gUser.ID),
-			zap.String("external.id", externalID),
-			zap.String("okta.user.id", u.Id),
-			zap.String("okta.user.email", email),
-		)
+			l.Debug("created governor user from okta sync",
+				zap.String("governor.user.id", gUser.ID),
+				zap.String("external.id", externalID),
+				zap.String("okta.user.id", u.Id),
+				zap.String("okta.user.email", email),
+			)
+		}
 
 		created++
 
 		return u, nil
 	}
 
-	l.Info("starting to sync missing okta users into governor")
+	l.Info("starting to sync missing okta users into governor", zap.Bool("dry-run", dryrun))
 
 	users, err := oc.ListUsersWithModifier(ctx, syncFunc, &query.Params{})
 	if err != nil {
@@ -178,8 +187,10 @@ func syncUsersToGovernor(ctx context.Context) error {
 
 // deleteOrphanGovernorUsers is a helper function to delete governor users that not longer exist in okta
 func deleteOrphanGovernorUsers(ctx context.Context, gc *governor.Client, extIDMap map[string]struct{}) (int, error) {
+	dryrun := viper.GetBool("sync.dryrun")
+
 	l := logger.Desugar()
-	l.Info("starting to clean orphan governor users")
+	l.Info("starting to clean orphan governor users", zap.Bool("dry-run", dryrun))
 
 	var deleted int
 
@@ -219,8 +230,10 @@ func deleteOrphanGovernorUsers(ctx context.Context, gc *governor.Client, extIDMa
 			zap.String("governor.user.email", gu.Email),
 		)
 
-		if err := gc.DeleteUser(ctx, gu.ID); err != nil {
-			return deleted, err
+		if !dryrun {
+			if err := gc.DeleteUser(ctx, gu.ID); err != nil {
+				return deleted, err
+			}
 		}
 
 		deleted++
@@ -250,25 +263,14 @@ func uniqueExternalIDs(users []*okt.User) map[string]struct{} {
 			continue
 		}
 
-		if extID == "" {
-			l.Error("empty external id from okta user",
-				zap.Error(err),
-				zap.String("okta.user.id", u.Id),
-			)
-
-			continue
-		}
-
-		e := fmt.Sprintf("ping|%s", extID)
-
-		if _, ok := extIDs[e]; ok {
+		if _, ok := extIDs[extID]; ok {
 			l.Warn("external id already exists in list of external ids",
 				zap.String("okta.user.id", u.Id),
-				zap.String("okta.user.external_id", e),
+				zap.String("okta.user.external_id", extID),
 			)
 		}
 
-		extIDs[e] = struct{}{}
+		extIDs[extID] = struct{}{}
 	}
 
 	l.Debug("returning list of unique external ids from okta users",
@@ -298,24 +300,13 @@ func email(u *okt.User) (string, error) {
 	return "", fmt.Errorf("email not found for user %s", u.Id) //nolint:goerr113
 }
 
-// externalID parses the externalID (pingSubject) from the okta user profile
+// externalID returns the id for the okta user
 func externalID(u *okt.User) (string, error) {
-	l := logger.Desugar()
-
-	// get the external id from the user profile
-	for k, v := range *u.Profile {
-		if k == "pingSubject" {
-			if pv, ok := v.(string); ok {
-				return pv, nil
-			}
-
-			l.Warn("okta user pingSubject in profile is not a string", zap.String("okta.user.id", u.Id), zap.Any("pingSubject", v))
-
-			return "", ErrOktaUserExternalIDNotString
-		}
+	if u.Id == "" {
+		return "", ErrOktaUserIDEmpty
 	}
 
-	return "", fmt.Errorf("external id not found for user %s", u.Id) //nolint:goerr113
+	return u.Id, nil
 }
 
 // userType parses the userType from the okta user profile
